PMP-5747 Fix EISDIR on stat directory

+++ android/src/main/java/com/rnfs/DownloadParams.java
@@ -12,11 +12,11 @@ public class DownloadParams {
   }
 
   public interface OnDownloadBegin {
-    void onDownloadBegin(int statusCode, long contentLength, Map<String, String> headers);
+    void onDownloadBegin(int statusCode, int contentLength, Map<String, String> headers);
   }
 
   public interface OnDownloadProgress {
-    void onDownloadProgress(long contentLength, long bytesWritten);
+    void onDownloadProgress(int contentLength, int bytesWritten);
   }
 
   public URL src;
